requirements.txt
flask==3.1.0
groq==0.25.0
notion-client==2.3.0
requests==2.32.3
gunicorn==23.0.0
duckduckgo-search==7.5.0
icalendar==6.1.3
python-dateutil==2.9.0
pdfplumber==0.11.6
fpdf2==2.8.3
caldav==1.4.0
anthropic==0.51.0
